CSC660 homework 1. I have tried to control the input stream a little more by reading it in byte by byte and building out the json string to be parsed into the Message object. This has had limited success. A lot of my work was wiped out (I think I need a tutorial on eclipse + git). Adding what I have so far. I think we need to look at locking down the input streams better as I still see a lot of data corruption. Also, fixed the stack trace spam by adding in a boolean that gets set to false when entering in the exception catch portion.

+++ Program1/ThreadedClient/src/edu/nku/common/LogicalClock.java
@@ -1,7 +1,7 @@
 package edu.nku.common;
 
 public class LogicalClock {
-	private int currentTime;
+	private double currentTime;
 	
 	public LogicalClock(){
 		currentTime = 0;
@@ -11,7 +11,7 @@ public class LogicalClock {
 		currentTime++;
 	}
 	
-	public synchronized void incrementTimeBySleep(int sleepTime){
+	public synchronized void incrementTimeBySleep(double sleepTime){
 		currentTime += sleepTime;
 	}
 	
@@ -19,7 +19,7 @@ public class LogicalClock {
 		currentTime = Math.max(currentTime + 1, otherTime + 1);
 	}
 	
-	public synchronized int getCurrentTime(){
+	public synchronized double getCurrentTime(){
 		return currentTime;
 	}
